share provider sources adding login in one class
